added comments (pt.1)

+++ 5-3.c
@@ -1,13 +1,14 @@
 #include <stdio.h>
 
 int main() {
-    double length, breadth;
+    // take length and breadth and input
+    float length, breadth;
     printf("Enter the length of the rectangle: ");
-    scanf("%lf", &length);
-
+    scanf("%f", &length);
     printf("Enter the breadth of the rectangle: ");
-    scanf("%lf", &breadth);
+    scanf("%f", &breadth);
 
+    // calculate area and perimeter, then display both
     printf("Area: %.2lf\nPerimeter: %.2lf\n", length * breadth, 2 * (length + breadth));
     return 0;
 }
